fix(netsuite): handle range issue (#4366)
## Context

Netsuite does not allow any value to be larger than 9 999 999.

## Description

For AI companies, when huge amount of tokens is consumed, `units` on fee
object can be really large. We will avoid such issue by setting quantity
to `1` and setting unit price to be the same as fee amount, only in such
edge cases.