feat(taxes): Soft-delete Tax model (#4370)
ISSUE-1142

**Make `Tax` model soft deletable.**

Currently, we have Taxes and we join the tax entries to other models via
many dedicated `AppliedTax` models.

They are 2 types of AppliedTaxes:
1️⃣  For invoice and fees, it's meant to be immutable
2️⃣ For plan, customers... it's meant to be a link that should only
exist if the tax exists

When a tax is "removed", it's currently hard deleted. 
- For AppliedTax 1️⃣ It's fine because the AppliedTaxes have a copy of
the name of the tax, the rate and so on. But the `tax_id` column is nil
instead of return the former primary key.
- For AppliedTa 2️⃣ , we delete them with the Tax.

This Fix will soft delete the tax and make sure AppliedTax 1️⃣ will
continue to retrieve the deleted tax.
AppliedTax 2️⃣ are still hard deleted, we want to this relationship to
go away.



When retrieving an invoice, the applied_taxes will continue to contain
the `lago_id`.

```diff
{
  "lago_id": "e1f55889-f136-4a1d-a951-5fbe754b5e93",
  "lago_invoice_id": "3e4bbbba-960c-42d6-91b7-df95ac0709af",
- "lago_tax_id": null,
+ "lago_tax_id": "59a9a21b-35fa-4f5a-836d-daa20e688012",
  "tax_name": "VAT",
  "tax_code": "vat-282e2794-264d-4ca7-9d6b-341be44372eb",
  "tax_rate": 20.0,
  "tax_description": "French Standard VAT",
  "amount_cents": 200,
  "amount_currency": "EUR",
  "fees_amount_cents": 0,
  "created_at": "2025-09-22T08:45:05Z"
}

```