chore(stripe): Remove "beta" for Stripe pre-auth feature (#4387)
## Context

This feature was introduced in March in #3262 as a _beta_ feature.

## Description

The _beta_ aspect was handled by a premium integration. I'm removing the
premium integration to release the feature in _GA_.

Note that "removing" integration isn't something too common, all apps
with this integration would be considered invalid.
I propose to introduce a list of previous integrations. 

Another way is a db migration to remove the data but it creates "down
time" where the feature is removed from DB but the code still rely on
this flag. I think this can be done later, if the list grows.