fix(event-aggregation): the query should group prorated units by day (#4399)
## Context

when for uniq count we post several events per day like:
27th June - 10:00 - property - Add
27th June - 11:00 - property - Remove
27th June - 12:00 - property - Add
27th June - 13:00 - property - Remove

current logic will count it as two events with minimal length of
proration 1 day, so the customer will be charged for 2 prorated days of
usage.

Instead, we want to group events by day, because we prorate them through
day, and for one day record only one event

another scenario:
27th June - 10:00 - property - Add
28th June - 10:00 - property - Remove
28th June - 11:00 - property - Add

if we "end" usage on 28th, the first Add event will "cover" usage on
28th. The event received on 28th will not do the double charge on 28th,
and will "start" on 29th, but: as result anyway we'll charge for the
usage on 27th, 28, 29.. so we can just ignore the remove event, and
aggregate all this as as one event started on 27th

## Description

Updated prorated uniq_count_queries for postgresql and clickhouse
Algorithm:
1) prepare events table
2) ignore only REMOVE events that will be "cancelled" by the following
Add event at the same day
3) using adjusted value logic nullify not needed Adds (if previous row
has the same operation_type, current has adjusted value 0; otherwise
it's 1 for Add; -1 for Remove)


this pr is a pure pain.
especially clickhouse - it can't do recursion, LAG and LOG are very
heavy there and tests were killing the db... so I needed to write a
completely different approach from postgres, thenI rewrote couple of
approaches on both stores, finally found a bug in clickhouse proration,
and here is the result