feat(customer): Add more filters to `customers` GraphQL query (#4404)
## Context

The filtering on the `GET /api/v1/customers` endpoint was too limited so we added a few filters:

- https://github.com/getlago/lago-api/commit/4a4a5ea
- https://github.com/getlago/lago-api/commit/2fcb515
- https://github.com/getlago/lago-api/commit/a1a8e37
- https://github.com/getlago/lago-api/commit/9bba4a9
- https://github.com/getlago/lago-api/commit/e04adee
- https://github.com/getlago/lago-api/commit/2f6064f

## Description

This adds those filters to the `customers` GraphQL query as well.

Note that GraphQL doesn't allow arbitrary map/hash as type so I had to create a specific type for the `metadata` filter.