fix(daily_usage): Fix backfill max timestamp with pro-ration (#4435)
## Context

This PR fixes the logic to back-fill the daily usage history. 

The existing implementation was forcing the `to_datetime` boundary in
the event stores to the day it was backfilling for. It was working fine
for standard metrics, but it was failling for all metrics relying on
proration as it was virtually reducing the duration of the period.

## Description

This PR introduces a new `max_timestamp` attribute in the event stores
and use it only for event filtering. It keeps relying on the
`to_datetime` attribute for all pro-rated computation making sure the
period duration is consistent on all consecutive daily usages