Remove customer snapshot from ReST API and GraphQL (#4448)
## Description:

This PR reverts the exposure of the customer_snapshot data from our REST
and GraphQL APIs.

## Context:

The associated feature was introduced to solve a live issue by ensuring
invoice data remains immutable.

However, as correctly flagged by @jdenquin , the current implementation
carries significant performance and scalability risks, particularly
concerning the data migration required for existing invoices. The change
was also merged without a formal design review, which is necessary for a
change of this magnitude.

To prevent usage of this unvetted feature we are removing it from all
public-facing APIs until a safer, more scalable solution is designed and
approved.

Changes in this PR:

* Removes the customer_snapshot association from the Invoice serializer
for the REST API.

* Removes the customer_snapshot field from the Invoice type in the
GraphQL API.

Out of Scope:

This PR does not drop the customer_snapshots table or the
CustomerSnapshot model. This will be handled in a subsequent migration
to ensure a safe rollback process.