Feat(fixed_charges-16): generate fees (#4348)
## Context

Add a service that generates fees for a fixed_charge. It uses calls
aggregation to see how much units should be billed, and then applies
charge model on the aggregation result, and then creates fee. By default
it does not apply taxes, but a flag can be passed to apply taxes

## Description

- added `Fees::FixedChargeService`