test(event): Use a `shared_example` for Postgres and Clickhouse event stores (#4452)
## Context

While working on fix an issue with the Clickhouse event store, I refactored the Clickhouse event store tests to make it easier to review and test. This created a discrepancy between the Postgres and Clickhouse test setup.

## Description

This extracts the updated tests into a `shared_example` and re-use it for Postgres.

Note that this change showcased two minor issues:

1. Since we're running tests in transactions and the events are created in the `events` connection, they were not available through `ActiveRecord::Base.connection.select_one`. This was replaced by `Event.connection.select_one`.
2. The Postgres store returns grouped result with a timestamp as a `Time` object and the Clickhouse store returns it as `String`.