feat(live_agg): Plug aggregated clickhouse store (#4395)
## Context

This PR is part of the Pre-aggregation epic, it follows
https://github.com/getlago/lago-api/pull/4277.

The main goal of this feature is to setup a complete aggregation
pipeline that will be leveraged to compute the customer usage without
re-querying the full set of events.

## Description

This PR refactor the `Events::Store::StoreFactory` to:
- Add a `new_instance` method, to respect the factory design pattern
- Accept a new `current_usage` flag on the
`Events::Store::StoreFactory#store_class`
- Check for the presence of the `clickhouse_live_aggregation` premium
feature on the organization when initializing the event store, to
instantiate a `Events::Store::AggregatedClickhouseStore` in case of
current_usage.

This will allow to enable the live aggregation on a per organization
basis in order to gradually validate the complete flow without impacting
the production