feat(multiple-payment-methods): add payment method model and table (#4455)
## Context

Currently in Lago, there can be only one payment provider customer and
also only one payment method.

## Description

With this feature, it will be possible to add multiple provider
customers and also attach many payment methods to it.

This is the first PR for this feature, that adds `payment_methods` table
and model