feat: Introduce middleware pattern to base service (#4462)
## Context

This PR is part of the live-aggregation epic.

The main goal of this feature is to setup a complete aggregation
pipeline that will be leveraged to compute the customer usage without
re-querying the full set of events.

Now that the aggregation pipeline is in place, we need a new system to
refresh the data when a plan, a charge or a filter gets updated.

This has to be done by monitoring the `Plans::CreateService` and
`Plans::UpdateService`. Since we need to compare results before and
after update, it has to be done "around" the existing `call` method and
as such requires a logic to be put in place.

As the approach is similar to the ActivityLog monitoring that is already
defined in the `BaseService`, the idea is to avoid "polluting" this
service with new logic. Here comes the middleware pattern!

## Description

This PR introduces a new `BaseMiddleware` class that can be used to
create "generic" logic that can be plugged in any class inheriting from
the `BaseService`.

For example, let's imagine that you want to add a logic to measure the
execution time of a service:
```ruby
class DurationMiddleware < BaseMiddleWare
  def before_call
     @start_time = Time.current
  end

  def after_call
     duration = Time.current - @start_time
     Rails.logger.info("#{kwargs[:label]} - #{service_instance.class} execution took #{duration} seconds")
  end
end
```

Alternatively you could also re-define the `call` method:
```ruby
class DurationMiddleware < BaseMiddleWare
  def call
     start_time = Time.current
    result = super
    
     duration = Time.current - start_time
     Rails.logger.info("#{service_instance.class} execution took #{duration} seconds")

    result
  end
end
```

To use in a service just use the `use` instruction:
```ruby
class MyDummyService < BaseService
  Result = BaseResult[:some_value]

  use DurationMiddleware, label: "[Test]" # example with arguments passed to the middleware

  def call
    result
  end
end
```

This new pattern has been used to refactor a bit the code base.
- The open telemetry tracer that was done in `BaseService#call` has been
extracted in the `Utils::LogTracerMiddleware` included by default in the
`BaseService`
- The ActivityLog monitoring has been moved into the
`UtilsActivityLogMiddleware` and is included when calling
`activity_loggable`

As a next step, a new `Plans::NotifyRefreshMiddleware` will be added to
track updates that need to be propagated in the pre-aggregated records

A new RSpec matcher has also been added to ease the test of middleware
inclusion:
```ruby
it { is_expected.to use_middleware(Utils::LogTracerMiddleware) }
# OR
it { expect(service_class).to use_middleware(Utils::LogTracerMiddleware) }
```