feat(live-agg): Add missing distinc_codes in preagg event store (#4497)
## Context

This PR is part of the Pre-aggregation epic

The main goal of this feature is to setup a complete aggregation
pipeline that will be leveraged to compute the customer usage without
re-querying the full set of events.

## Description

This PR add the missing `distinct_codes` method in the
`Events::Stores::AggregatedClickhouseStore`. It's goal is to avoid
performing events aggregation on billable metrics that did not received
any events in a period