fix: add small update on how prog billing credits are calculated (#4436)
## Context

Currently, if progressive billing amount that is billed during the
period is higher than the invoice charge fees, Lago will generate credit
note.

## Description

Credit note value that is described above is calculated as the
difference between progressive billing fees and final invoice fees. This
PR scopes invoice amount to the fees that are used in progressive
billing invoice.

There can be a case when charges are added during the period and current
calculation can lead to some inconsistencies. Customer should be
credited in the final invoice only for charges that are included in
progressive billing invoice, but currently customer could be credited
also for new charges. However, total balance is fine, it is just matter
how credits are spread between credit note and credits applied on the
final invoice