test(event): Move more event store tests to `shared_examples` (#4500)
## Context

https://github.com/getlago/lago-api/commit/b3131f433bfd38831a0a32c5deb1376659028902 introduced a `shared_examples` to improve consistency between the event stores tests. The following methods were not tested in the `shared_examples` though:

- `prorated_events_values`
- `prorated_sum`
- `grouped_prorated_sum`
- `weighted_sum`
- `grouped_weighted_sum`
- `prorated_unique_count`
- `grouped_prorated_unique_count`
- `prorated_unique_count_breakdown`

Furthermore, the `weighted_sum_breakdown` method was not tested.

## Description

This moves the following methods tests to the `shared_examples`:

- `prorated_events_values`
- `prorated_sum`
- `grouped_prorated_sum`
- `weighted_sum`
- `grouped_weighted_sum`
 
Tests were also added for the `weighted_sum_breakdown` method.

It looks like the behavior of the prorated unique count methods actually differs from the Postgres and Clickhouse stores so they were not moved to the `shared_examples`.

This also fixes an issue with the `weighted_sum_breakdown` method of the Clickhouse store which returned the `second_duration` as string rather than int. That's because `date_diff` actually returns an `Int64` and ActiveRecord transform that into a `String` so we call `#to_i` to ensure consistency between implementations:

```ruby
lago-api(staging)> Clickhouse::BaseRecord.connection.select_one("SELECT 1::Int64")
=> {"CAST('1', 'Int64')" => "1"}
lago-api(staging)> Clickhouse::BaseRecord.connection.select_one("SELECT 1::Int32")
=> {"CAST('1', 'Int32')" => 1}
```

Note that the `weighted_sum_breakdown` method of the Clickhouse stores returns timestamps with a 5 decimal precision rather than 3 as in the other methods.