chore(seeds): Update seeds to avoid events in future (#4512)
## Context

The seeds currently generate events with timestamp `(offset.month - rand(1..20).days).ago` when `offset` goes from `0` to `5`. This means that for offset `0`, the timestamp will be in the future. I believe this is a bug introduced in https://github.com/getlago/lago-api/commit/7aa074069298da6c094b26447cd3669995105e8d as we previously defined the timestamp as `Time.current - offset.month - rand(1..20).days`.

## Description

This changes the events creation to ensure all events are in the past.