fix(self-hosted): Make sure that event_id removal on cached_aggregation is safe (#4514)
## Context

This PR follows https://github.com/getlago/lago-api/pull/4506

The removal of the `cached_aggregations.event_id` column was not made in
a safe way as it was removing the column and the code relying on it in
the same migration.

## Description

This PR reworks the approach to make sure it can be released for self
hosted instances in a safe way.

It:
- "void" the database migration removing the column
- Add the event_id column in the "ignored" list on the
`CachedAggregation` model

A new PR will come later in a new release to finally remove the column
(if it exists)