chore(cache): Ensure cache store DB consistency in dev environment (#4526)
## Context

https://github.com/getlago/lago/commit/3cd78f1cdfd94dda9ea0c248bd074d8a20f868a7 fixed the `events-processor` cache Redis DB in development environment. But if someone were to change the `.env.development.default`, we would again have discrepancy between the `events-processor` and `lago-api` cache Redis DB.

## Description

This ensure we have consistency between the two services in development environment by relying on the `LAGO_REDIS_CACHE_DB` env variable for `lago-api` as well and defaulting to the DB `0`.

Note that this could cause the cache to switch to DB `0` if the `lago` repository is not up-to-date with `main`. This is acceptable in development environment.