feat(integration): Raise BadGatewayError on 502 errors (#4494)
If Anrok returns a Bad Gateway error with a 200 status, we handle this
by raising a custom `Integrations::Aggregator::BadGatewayError`. See
https://github.com/getlago/lago-api/pull/3234

This error can be use to retry the job, like for
RefreshOngoingBalanceJob

https://github.com/getlago/lago-api/blob/b1596cd6af82b0e9871949b5fd25438732bb2e49/app/jobs/wallets/refresh_ongoing_balance_job.rb#L11

Unfortunately, of the error is a return with a 502 status, we'll raise a
regular `LagoHttpClient::HttpError`, which is not retrie

<img width="2456" height="612" alt="CleanShot 2025-10-15 at 15 50 43@2x"
src="https://github.com/user-attachments/assets/0ded4d2c-0bd9-40e7-be44-990ac735463f"
/>



This PR aligns the behavior to always raise this custom error.

Notice that this custom error extends the HttpError:

https://github.com/getlago/lago-api/blob/b1596cd6af82b0e9871949b5fd25438732bb2e49/app/services/integrations/aggregator/bad_gateway_error.rb#L5


## Main changes

1. refactor tests to use webmock instead of mocking the client
2. rename `BAD_REQUEST_ERROR` to `BAD_GATEWAY_ERROR`
3. make bad_gateway_error?(e) return true if status is 502 (not just
content grepping)
4. raise `BadGatewayError` instead of `HttpError`