feat(subscription): Expose subscription_at in milliseconds (#4521)
When retrieving events to be billed, we'll only consider events sent
after the `subscription.started_at`.

On Subscription model, we store `started_at` with nanoseconds precision:
`2024-04-23T10:01:09.123456Z`
On Event model in CH, we store the `timestamp` at the millisecond
precision: `2024-04-23T10:01:09.123Z`
On Event model in PG, we store the `timestamp` at the nanoseconds
precision: `2024-04-23T10:01:09.123456Z`, but in the service, converting
the timestamp to a ruby datetime will lose some precision.

This means for a subscription started at 10:01:09.666, an event sent at
10:01:09.000 will not be counted. This is not a bug, nor recent, but our
API only expose the `started_at` at the second precision so this is
impossible to know if an event at the same second will be counted or
not.

**Adding precision is not a breaking change because this is still
`iso8601` format**, but with more precision.

This PR adds precision down to the microseconds (3 places).