feat(api): Introduce `timestamp_from_started_at` filters for Event List endpoint (#4531)
When billing a customer, we only count events where timestamps is after
the `started_at`.

In the API, to retrieve only valid events we introduce this new boolean
filter instead of requesting the precise timestamp.

In a following PR, I'll add validation on the filters as well. This
filter can only be passed if `timestamp_from` is not set. I think the
cast to bool will be done in the validation as well.

In term of performance, there is nothing new, we can already filter on
timestamps so the indexes already exists. The one thing added is
retrieving a subscription by external id.