feat(estimate): Prepare Clickhouse event estimate (#4542)
## Context

Organizations configured to use the Clickhouse event store are not able
to use the `POST /api/v1/events/estimate_fees` endpoint as the logic
relies on the creation of a temporary record in the `events` table of
Postgres.

This PR is part of a refactor allowing the estimate with clickhouse
event stores and speeding up te estimate computation with postgres store
by removing the need for persisting a temporary record.

## Description

Since the change requires a lot of file update, it will be split in
multiple parts.
This first part:
- Adds a new `persisted` attribute to the `Events::Common` structure,
with a default value to true. It will allow us to mark define an event
as "temporary"
- Define the `==` method on the `BaseResult` class, to allow comparison
of result states and values in specs
- Extract the charge model class initialization for in advance billing
from `Charges::ApplyPayInAdvanceChargeModelService` to
`ChargeModels::Factory`