fix(tax): Silent errors with parallel invoice refresh (#4552)
## Description

The goal of this PR is to silent `ActiveRecord::InvalidForeignKey`
raised from `Invoices::ProvidersTaxes::PullTaxesAndApplyJob` when a
draft invoice is being refreshed in parallel of the tax process.
The error occurs because the fees are removed by the refresh process,
and as such, the service is unable to persist the fees_applied_taxes.

The service keeps raising if the invoice is not in draft to avoid
loosing other potential errors