feat(estimate_fee): Add spec scenarios (#4555)
## Context

Organizations configured to use the Clickhouse event store are not able
to use the `POST /api/v1/events/estimate_fees` endpoint as the logic
relies on the creation of a temporary record in the `events` table of
Postgres.

This PR is part of a refactor allowing the estimate with clickhouse
event stores and speeding up te estimate computation with postgres store
by removing the need for persisting a temporary record.

## Description

- This PR is adding spec scenarios covering as much use case as possible
of the estimate endpoint. It will ease the refactor and the clickhouse
support implementation by reducing the risk of breaking changes.

The support of the dynamic charge model is also added to the estimate
endpoint as it was missing.