Feat/e invoices invoices (#4541)
This PR is part of the e-invoicing set of pull requests.

It has the necessary code to generate the files for Invoices.
It depends on https://github.com/getlago/lago-api/pull/4529 to be merged
first because FacturX requires `pdfcpu` to embed XML into pdf documents.

Please review it looking commit by commit, it's going to be easier to
follow (_trust me, I promise_).