fix(credit_notes): Take rouding into account with credit notes validation (#4553)
## Context

This PR is related to the recent changes around taxes and rounding on
credit notes:
- https://github.com/getlago/lago-api/pull/4373
-  https://github.com/getlago/lago-api/pull/4411
- https://github.com/getlago/lago-api/pull/4449

Some credit notes created before the fix might be affected by a
validation issue when the `credit_amount` is 1 cent above max creditable
invoice value (due to past rounding issues). This could also potentially
affect new consecutive credit notes emitted on an invoice because of the
+/- that we are applying on the total/credit_amount

## Description

This fix updates the validations so that the `credit_amount` and
`total_amount` are not considered invalid if they are 1 cents at most
above the max creditable amount.