misc(estimate): Add persentage scenario (#4572)
## Context

Organizations configured to use the Clickhouse event store are not able
to use the `POST /api/v1/events/estimate_fees` endpoint as the logic
relies on the creation of a temporary record in the `events` table of
Postgres.

This PR is part of a refactor allowing the estimate with clickhouse
event stores and speeding up te estimate computation with postgres store
by removing the need for persisting a temporary record.

## Description

This PR follows https://github.com/getlago/lago-api/pull/4555
It adds one more scenario to cover the sum aggregation with percentage
charge model including free units per transaction.
It also fixes one case with clickhouse when using dynamic charge model
if no events are found in the period