chore(cache): Reduce API key cache TTL in development environment (#4573)
## Context

To optimize performance, we cache API key along with the linked organization for 1 hour. In development environment though, we might often update the organization with premium integrations or enable a specific event store. This can lead the organization to be cached with a different value and therefore to API calls behaving differently from what we might expect.

## Description

This reduces the API key TTL to 10 seconds in development environment to ensure that:

- We still rely on the cache and test it properly locally
- Ensure that any organization update will be reflected at most 10 seconds later