fix(invoice-preview): include ending_at logic in invoice preview (#4576)
## Context

Currently invoice preview endpoint does not take into account
`ending_at` field.

## Description

This PR adds changes that handles described case. It is not necessary
anymore to pass `terminated_at`, but subscription is terminated
automatically (in the preview context only and without any perisisting)
if `ending_at` is falling in the current period.

With multiple subscriptions, we will handle it as before.