feat(estimate): Allow event estimate with Clickhouse (#4536)
## Context

Organizations configured to use the Clickhouse event store are not able
to use the `POST /api/v1/events/estimate_fees` endpoint as the logic
relies on the creation of a temporary record in the `events` table of
Postgres.

This PR is part of a refactor allowing the estimate with clickhouse
event stores and speeding up te estimate computation with postgres store
by removing the need for persisting a temporary record.

## Description

This PR is the last step of the feature. It removes the need for a
"fake" persisted event (and so the transaction opened) on the events
table to estimate the event.
For estimate, the logic is turned from a 
`result = amount_including_fake_persisted_event -
amount_without_event_value`
to
`result = amount_from_already_persisted_events +
amount_plus_event_value`