fix(event): Add back transaction_id presence validation (#4580)
## Context

`transaction_id` presence validation on event was removed by mistake
with https://github.com/getlago/lago-api/pull/4133.
The initial intention was only to remove the uniqueness validation.

## Description

This PR add back the presence validation to avoid `PG::NotNullViolation`
to be raised when trying to create an event without the transaction_id
value