fix(clickhouse): Force BigDecimal for weighted sum (#4581)
## Description

This PR makes sure that the weighted_sum and grouped_weighted_sum
computations in clickhouse always returns a `BigDecimal`

Under some unclear circumstances, it appears to be returning a string
leading to unexpected errors. It might be an error that is not properly
handled by the `clickhouse-activerecord` gem.
This will not fix the root cause, but it should at least gives us the
invalid value if it cannot be converted into a `BigDecimal`

NOTE: the same logic will be applied to other aggregation types