chore(seeds): Remove `field_name` from seeds' count billable metric (#4584)
## Context

The `count_bm` billable metric in the seeds currently has a `field_name` defined but `count_agg` billable metric should not have a `field_name`. This causes a visual bug on the UI.

Note that there's currently no validation ensuring that this field is `nil` when `aggregation_type` is `count_agg`.

## Description

This updates the seeds to remove the `field_name` and rely on the `BillableMetrics::CreateService`.