fix(webhooks): Improve Webhook and InboundWebhook cleanup tasks
Currently we have failing jobs for Webhook cleanup clock jobs, while the
error does not return any message back, just job failing, we see there
are a big number of webhooks that should have been deleted and are not.

This change makes deletion faster as "#destroy_all" was loading each
record in memory and running `#destroy`. However, `delete_all` runs a
single SQL query to delete them all.

Webhooks and InboundWebhooks does not have callbacks, nor associations
to be considered on deletion. Hence, this is safe and faster.