chore(sidekiq): Allow to configure Sidekiq concurrency in dev env (#4589)
## Context

When testing concurrency problematics locally, we would usually enable the different workers (`events`, `billing`, etc.). The current setting make it so that each worker would spawn 10 threads. This would result in 60+ Sidekiq threads which can overwhelm the Docker resources and be counter productive.

It also makes sense to configure those independently. For instance, when testing invoice concurrency, we wouldn't need to run 10 `pdf` or `webhook` workers.

## Description

This allows to define the concurrency dynamically. A later change to https://github.com/getlago/lago will then allow to define the concurrency for each worker independently.