feat(integration): Add `billing_entity_id` to integration collection mapping (#4445)
## Context

Currently, our customers cannot sync invoices to multiple accounting entities (subsidiaries) as our mapping is made at the organization level. However, in NetSuite, for instance, the tax and item mapping are made at the entity level. We need to allow our mapping to be defined at the entity level.

## Description

This is the first change where we add a `billing_entity_id` field to the `integration_collection_mappings` along with the proper indexes and validations.

There will be two follow-up change to:

1. Update the `integration_mappings` similarly.
2. Update the integration payload services to define a lookup and fallback strategy based on the available mappings.