test(integration): Update integration payloads tests (#4458)
## Context

Currently, our customers cannot sync invoices to multiple accounting
entities (subsidiaries) as our mapping is made at the organization
level. However, in NetSuite, for instance, the tax and item mapping are
made at the entity level. We need to allow our mapping to be defined at
the entity level.

## Description

This heavily tests the changes introduced in #4450 by relying on a
`shared_example`.