feat(integration): Add `billing_entity_id` to integration mapping (#4446)
## Context

Currently, our customers cannot sync invoices to multiple accounting entities (subsidiaries) as our mapping is made at the organization level. However, in NetSuite, for instance, the tax and item mapping are made at the entity level. We need to allow our mapping to be defined at the entity level.

## Description

This is a follow-up to https://github.com/getlago/lago-api/commit/bc300c7a3bddcc9e209b3212d4c6ded5301a9a1f.

This adds a `billing_entity_id` field to the `integration_mappings` along with the proper indexes and validations.

There will be a follow-up change which updates the integration payload services to define a lookup and fallback strategy based on the available mappings.