feat(issuing_date-1): Add issuing date preference settings to BillingEntity and Customer (#4595)
## Roadmap Task

👉
[Spec](https://www.notion.so/getlago/Spec-Define-the-issuing-date-preferences-of-subscription-invoices-2a0ef63110d2807d860cc22ced334bfb)
👉 [Dive
In](https://www.notion.so/getlago/BE-Dive-In-Define-the-issuing-date-preferences-of-subscription-invoices-2a3ef63110d2800e8620fb5687f0edf9?d=2a3ef63110d2808f96fd001cbc71298b#2a3ef63110d2807abf14f498b984ac57)

## Context

Lago, since the beginning, creates the invoices with an issuing date at
the moment the invoice is issued. We got multiple feedback from
customers, where, in that case, they would like the invoice
`issuing_date` to be on the last day of the billing period

## Description

This PR is the first part – it adds issuing date settings to both
`BillingEntity` and `Customer`.